requirements.txt
paddleocr
paddlepaddle
opencv-python
numpy
python-dotenv
mysql-connector-python
setuptools
